refactor(senpi-task): extract the RPC model catalog probe transport
Move the spawn/stream/parse half of process model admission into its own
module so the admission file keeps room for the catalog policy itself. The
public surface is unchanged: model-admission.ts re-exports the probe symbols
that packages/omo-senpi and the RPC tests already import.
